complete feature-base

+++ src/main/java/com/example/managershop/entity/IdRoleUser.java
@@ -5,5 +5,5 @@ import java.io.Serializable;
 @SuppressWarnings("unused")
 public class IdRoleUser implements Serializable {
     private Long idUser;
-    private Long idRole;
+    private int idRole;
 }
